requirements.txt
Flask>=2.0
python-nmap
fpdf2
tqdm
python-dotenv
